+++ CHANGELOG.md
@@ -4,7 +4,7 @@
   * ...
 
 3.x.x - 2024xxxx
-  * ...
+  * Attach notes db instead of opening new connection
 
 3.5.0 - 20241025 Expand metadata parsing + add models
   * Add traits + loading of authors, series and notes